--- projects/components-demo/src/app/demo-select/demo-select.component.html
@@ -1,19 +1,0 @@
-<div class="demo-wrapper">
-  <div class="demo-title">Select</div>
-
-  <div class="demo-border">
-    <label class="demo-subtitle">Numbers</label>
-    <dp-select placeholder="Please select" [formControl]="formControl" [options]="options"></dp-select>
-    <div class="demo-errors" *ngIf="errors">{{errors}}</div>
-
-    <div style="margin-top: 1rem; margin-bottom: 1rem;">
-      <dp-button color="primary" *ngIf="this.formControl.disabled" (click)="enable()">Enable</dp-button>
-      <dp-button color="primary" *ngIf="!this.formControl.disabled" (click)="disable()">Disable</dp-button>
-      <dp-button color="secondary" (click)="error()">Error</dp-button>
-    </div>
-
-    <div class="demo-value">
-      Value: {{formControl.value}}
-    </div>
-  </div>
-</div>
