+++ projects/components-demo/src/app/demos/demo-select/demo-select.component.html
@@ -1,38 +1,40 @@
 <div class="component-intro">
-  <h1>Multiselect Input</h1>
+  <h1>Select Input</h1>
   <p>
-    A selectable input component for choosing multiple values from a drop down menu. Utilizes the same component as the
-    <a routerLink="/select"><code>SelectComponent</code></a>. Supports <code>ControlValueAccessor</code>.
+    A selectable input component for choosing a single value from a drop down menu.
+    Supports <code>ControlValueAccessor</code>.
   </p>
 </div>
 
 <div class="demo-wrapper">
-    <h4>Numbers</h4>
-    <dp-select placeholder="Please select" [formControl]="formControl" [options]="options" [multiselect]="true"></dp-select>
-    <div class="demo-errors" *ngIf="errors">{{ errors }}</div>
+  <h4>Numbers</h4>
+  <dp-select placeholder="Please select" [formControl]="formControl" [options]="options"></dp-select>
+  <div class="demo-errors" *ngIf="errors">{{ errors }}</div>
 
-    <div class="demo-btn-group">
-      <dp-button color="primary" *ngIf="this.formControl.disabled" (click)="enable()">Enable</dp-button>
-      <dp-button color="primary" *ngIf="!this.formControl.disabled" (click)="disable()">Disable</dp-button>
-      <dp-button color="secondary" (click)="error()">Error</dp-button>
-    </div>
+  <div class="demo-btn-group">
+    <dp-button color="primary" *ngIf="this.formControl.disabled" (click)="enable()">Enable</dp-button>
+    <dp-button color="primary" *ngIf="!this.formControl.disabled" (click)="disable()">Disable</dp-button>
+    <dp-button color="secondary" (click)="error()">Error</dp-button>
+  </div>
 
-    <div class="demo-value">Value: {{ formControl.value }}</div>
+  <div class="demo-value">Value: {{ formControl.value }}</div>
 </div>
 
 <div class="component-info">
   <dp-tab-group>
     <dp-tab label="Documentation">
       <h3>Getting Started</h3>
-      <p>To enable multi-select in the <code>SelectComponent</code>, add the <code>[multiselect]="true"</code> attribute.
-        The multi-select component can be used with Angular Form Control.</p>
-      <app-codeblock>{{ codeGetStarted }}</app-codeblock>
+      <p>
+        The select component requires a list of options to create an <code>SelectItemModel</code> with. These options
+        consist of a label and a value. Multi-selection can also be enabled with the <code>SelectComponent</code>,
+        details can be found in the <a routerLink="/multi-select">Multiselect Input</a> demo.
+      </p>
       <h4>Import</h4>
       <app-codeblock>Currently everything is imported by DevproComponentsModule.</app-codeblock>
 
       <h3>Options</h3>
       <p>The values contained the <code>SelectComponent</code>'s drop down can be easily passed in with the <code>options</code>
-      attribute. The options can be defined in a typescript file.</p>
+        attribute. The options can be defined in a typescript file.</p>
       <app-codeblock>{{ codeOptionsAttr }}</app-codeblock>
       <app-codeblock>{{ codeOptions }}</app-codeblock>
 
@@ -66,7 +68,9 @@
           <td>disabled</td>
           <td>boolean</td>
           <td>false</td>
-          <td>Disabled attribute of the select component, when true it will visibly change the input field and make it un-clickable.</td>
+          <td>Disabled attribute of the select component, when true it will visibly change the input field and make it
+            un-clickable.
+          </td>
         </tr>
       </table>
