+++ projects/components-demo/src/app/demos/demo-select/demo-select.component.scss
@@ -3,3 +3,8 @@ dp-button {
   margin-top: 0.5rem;
   margin-right: 0.5rem;
 }
+
+a {
+  color: rgba(59, 130, 246, 1);
+  text-decoration: none;
+}
